+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix binstubs not being replaced when their quoting style was changed (#534)
 * Preserve comments right after the shebang line which might include magic comments such as `frozen_string_literal: true'
+* Added support for `gems.rb` with Gemfile file name detection using Bundler method
 
 ## 2.0.2